+++ recpack/algorithms/mult_vae.py
@@ -18,8 +18,7 @@ import numpy as np
 
 from recpack.algorithms.base import TorchMLAlgorithm
 from recpack.algorithms.loss_functions import vae_loss
-from recpack.algorithms.util import naive_sparse2tensor
-from recpack.scenarios.splitters import yield_batches
+from recpack.algorithms.util import naive_sparse2tensor, get_batches
 
 logger = logging.getLogger("recpack")
 
@@ -187,7 +186,7 @@ class MultVAE(TorchMLAlgorithm):
 
         np.random.shuffle(users)
 
-        for batch_idx, user_batch in enumerate(yield_batches(users, self.batch_size)):
+        for batch_idx, user_batch in enumerate(get_batches(users, self.batch_size)):
             X = naive_sparse2tensor(train_data[user_batch, :]).to(self.device)
 
             # Clear gradients